Start raising real exceptions
Currently the code tries to raise exceptions with raise_swig_error.
This is duplicated between the python and perl .i files and, at least in
python it doesn't work. The exception is not passed to the interpreter
at all.

Removing the method, and substituting its call to the new
throw_exception defined in except.c.

Signed-off-by: Guido Trotter <ultrotter@quaqua.net>