Move constants to nfq_constants.h
This header file is used so we can import constants from external
headers, without copying the numbers manually. The netfilter.h constants
are moved there, and new constants for libnfqueue copy modes are added.

Const ints are used because re-defining a variable as itself, or
defining it as an enum value isn't working.

Signed-off-by: Guido Trotter <ultrotter@quaqua.net>